+++ lib/Context/ThemeContext.js
@@ -6,7 +6,7 @@ import {
   useCallback,
 } from "react";
 import { ThemeProvider } from "styled-components";
-import theme from "@/components/ThemePicker";
+import theme from "@/lib/Context/ThemePicker";
 import { animate, useMotionValue } from "framer-motion";
 
 const StateContext = createContext();
